refactor: move code to routes
